test(#664): extract completions --install subprocess to integration; ALLOWED_SPAWN -> 0 (C6.1-spawn)
completions.test.ts was already mostly in-process; only the `--install` test
genuinely needed a subprocess (it asserts the install-path message warn() emits
to stderr, which runCliCapture cannot surface under the suite preload). Extract
that one test into tests/integration/completions-install.test.ts and drop the
now-unused spawnSync/fs/path imports from the unit file, leaving it spawn-free.

This empties ALLOWED_SPAWN entirely (real-spawn lane of the unit-purity ratchet
is now 0). UNIT_PURITY_BASELINE 14 -> 13 (only the full-index lane remains).
