+++ tests/completions.test.ts
@@ -1,7 +1,4 @@
 import { afterAll, beforeAll, describe, expect, test } from "bun:test";
-import { spawnSync } from "node:child_process";
-import fs from "node:fs";
-import path from "node:path";
 import { runCliCapture } from "./_helpers/cli";
 import { makeSandboxDir, type SandboxedDir, withEnv } from "./_helpers/sandbox";
 
@@ -138,38 +135,10 @@ describe("completions command", () => {
   });
 });
 
-// ── Integration: --install ───────────────────────────────────────────────────
-
-describe("completions --install", () => {
-  // KEPT AS A SUBPROCESS: this asserts the install-path message emitted via
-  // warn() → stderr, which the in-process harness does not surface under the
-  // suite-wide test preload. A real subprocess exercises the user-visible stderr.
-  test("writes completion file to XDG_DATA_HOME path", () => {
-    const xdgData = makeTempDir();
-    const result = spawnSync("bun", ["./src/cli.ts", "completions", "--install"], {
-      encoding: "utf8",
-      timeout: 10_000,
-      cwd: path.resolve(import.meta.dir, ".."),
-      env: {
-        ...process.env,
-        AKM_STASH_DIR: undefined,
-        HOME: isolatedHome,
-        XDG_CACHE_HOME: xdgCache,
-        XDG_CONFIG_HOME: xdgConfig,
-        XDG_DATA_HOME: xdgData,
-      },
-    });
-
-    expect(result.status).toBe(0);
-    const expectedPath = path.join(xdgData, "bash-completion", "completions", "akm");
-    expect(result.stderr).toContain(expectedPath);
-    expect(fs.existsSync(expectedPath)).toBe(true);
-
-    const content = fs.readFileSync(expectedPath, "utf8");
-    expect(content).toStartWith("#!/bin/bash");
-    expect(content).toContain("complete -F _akm akm");
-  });
-});
+// The `completions --install` subprocess test lives in
+// tests/integration/completions-install.test.ts — it asserts the install-path
+// message emitted via warn() → stderr, which the in-process harness cannot
+// surface under the suite-wide test preload, so it needs a real subprocess.
 
 // ── Unsupported shell ────────────────────────────────────────────────────────
